Move debug utilities
